tech(api): move POST /api/assessments endpoint to evaluation BC
