+++ api/src/evaluation/application/assessments/assessment-routes.js
@@ -83,6 +83,27 @@ const register = async function (server) {
         ],
       },
     },
+    {
+      method: 'POST',
+      path: '/api/assessments',
+      config: {
+        auth: false,
+        validate: {
+          options: {
+            allowUnknown: true,
+          },
+          payload: Joi.object({
+            data: Joi.object({
+              attributes: Joi.object({
+                type: Joi.string().required(),
+              }).required(),
+            }).required(),
+          }).required(),
+        },
+        handler: assessmentController.save,
+        tags: ['api'],
+      },
+    },
   ];
   server.route(routes);
 };
