+++ api/tests/evaluation/unit/application/assessments/assessment-controller_test.js
@@ -8,10 +8,12 @@ import { stageUsecases } from '../../../../../src/prescription/stages/domain/use
 import { usecases as profileUsecases } from '../../../../../src/profile/domain/usecases/index.js';
 import { usecases as questUsecases } from '../../../../../src/quest/domain/usecases/index.js';
 import { DomainTransaction } from '../../../../../src/shared/domain/DomainTransaction.js';
+import { Assessment } from '../../../../../src/shared/domain/models/Assessment.js';
 import { featureToggles } from '../../../../../src/shared/infrastructure/feature-toggles/index.js';
 import { expect } from '../../../../test-helper.js';
+import { hFake } from '../../../../tooling/mocks/hapi.mock.js';
 
-describe('Unit | Controller | assessment-controller', function () {
+describe('Evaluation | Unit | Application | assessment-controller', function () {
   describe('#completeAssessment', function () {
     let assessmentId, assessment, locale;
 
@@ -234,6 +236,61 @@ describe('Unit | Controller | assessment-controller', function () {
         // then
         expect(questUsecases.getQuestResultsForCampaignParticipation).to.have.been.called;
         expect(profileUsecases.shareProfileReward).to.not.have.been.called;
+      });
+    });
+  });
+
+  describe('#save', function () {
+    context('when the assessment saved is a preview test', function () {
+      const request = {
+        headers: {
+          authorization: 'Bearer my-token',
+        },
+        payload: {
+          data: {
+            id: 42,
+            attributes: {
+              'estimated-level': 4,
+              'pix-score': 4,
+              type: 'PREVIEW',
+            },
+            relationships: {
+              course: {
+                data: {
+                  id: 'null-preview-id',
+                },
+              },
+            },
+          },
+        },
+      };
+      let assessmentRepositoryStub;
+
+      beforeEach(function () {
+        assessmentRepositoryStub = { save: sinon.stub() };
+        assessmentRepositoryStub.save.resolves({
+          toDto: () => {
+            return {};
+          },
+        });
+      });
+
+      it('should save an assessment with type PREVIEW', async function () {
+        // given
+        const expected = new Assessment({
+          id: 42,
+          courseId: null,
+          type: 'PREVIEW',
+          userId: null,
+          state: 'started',
+          method: 'CHOSEN',
+        });
+
+        // when
+        await assessmentController.save(request, hFake, { assessmentRepository: assessmentRepositoryStub });
+
+        // then
+        expect(assessmentRepositoryStub.save).to.have.been.calledWithExactly({ assessment: expected });
       });
     });
   });
